+++ CHANGES.rst
@@ -18,6 +18,7 @@ in progress
 - Add support for XLSX and ODS input formats
 - Add test case for basic nested JSON document
 - Add ``frictionless`` backend
+- Add support for Google Sheets input format
 
 
 2023-01-xx 0.0.0